add_protect_actions_and_delete_file_from_disc

+++ files.go
@@ -6,6 +6,7 @@ import (
 	"fmt"
 	"io/ioutil"
 	"net/http"
+	"os"
 	"path"
 	"strings"
 
@@ -38,7 +39,6 @@ type Attachment struct {
 	Title       string `json:"title"`
 	Description string `json:"description" gorm:"type:text"`
 	IsMain      int    `json:"isMain"`
-	Bind        string `json:"bind"`
 	Hash        string `json:"hash"`
 	Index       int    `json:"index" gorm:"type:int(6)"`
 }
@@ -53,9 +53,66 @@ func Configure(a core.App) {
 	//protect CRUD actions with files info
 	App.R.HandleFunc("/api/files", actionGetAll).Methods("GET")
 	App.R.HandleFunc("/api/files/{id}", actionGetOne).Methods("GET")
-	App.R.HandleFunc("/api/files", actionUpload).Methods("POST")
-	App.R.HandleFunc("/api/files/{id}", actionReUpload).Methods("PATCH")
-	App.R.HandleFunc("/api/files/{id}", actionDelete).Methods("DELETE")
+	App.R.HandleFunc(
+		"/api/files",
+		App.Protect(
+			actionUpload,
+			[]string{"admin", "user"})).Methods("POST")
+	App.R.HandleFunc(
+		"/api/files/{id}",
+		App.Protect(
+			actionReUpload,
+			[]string{"admin", "user"})).Methods("PATCH")
+	App.R.HandleFunc(
+		"/api/files/{id}",
+		App.Protect(
+			actionDelete,
+			[]string{"admin", "user"})).Methods("DELETE")
+}
+
+func upload(r *http.Request) (File, error) {
+	// Parse our multipart form, 10 << 20 specifies a maximum
+	// upload of 10 MB files.
+	r.ParseMultipartForm(10 << 20)
+	// FormFile returns the first file for the given key `myFile`
+	// it also returns the FileHeader so we can get the Filename,
+	// the Header and the size of the file
+	file, handler, err := r.FormFile("file")
+	if err != nil {
+		return File{}, errors.New("Error Retrieving the File")
+	}
+	defer file.Close()
+	filename := strings.TrimSuffix(handler.Filename, path.Ext(handler.Filename))
+	fileext := path.Ext(handler.Filename)
+
+	// Create a temporary file within our temp-images directory that follows
+	// a particular naming pattern
+	tempFile, err := ioutil.TempFile("web/uploads", "*"+filename+fileext)
+	if err != nil {
+		return File{}, err
+	}
+	defer tempFile.Close()
+
+	// read all of the contents of our uploaded file into a
+	// byte array
+	fileBytes, err := ioutil.ReadAll(file)
+	if err != nil {
+		return File{}, err
+	}
+	// write this byte array to our temporary file
+	tempFile.Write(fileBytes)
+
+	return File{
+		UserID: 0,
+		Name:   filename,
+		Path:   tempFile.Name(),
+		Ext:    fileext,
+		Preset: "notset",
+		Size:   handler.Size,
+		Status: 0,
+		Type:   0,
+		Hash:   fmt.Sprintf("%x", md5.Sum(fileBytes)),
+	}, nil
 }
 
 func actionGetAll(w http.ResponseWriter, r *http.Request) {
@@ -164,22 +221,28 @@ func actionReUpload(w http.ResponseWriter, r *http.Request) {
 		rsp       = core.Response{Data: &filemodel}
 	)
 
-	data, err := upload(r)
-	if err != nil {
-		rsp.Errors.Add("file", err.Error())
-	} else {
-		vars := mux.Vars(r)
-		App.DB.First(&filemodel, vars["id"])
+	vars := mux.Vars(r)
+	App.DB.First(&filemodel, vars["id"])
 
-		if filemodel.ID == 0 {
-			rsp.Errors.Add("ID", "Contentelement not found")
+	if filemodel.ID == 0 {
+		rsp.Errors.Add("ID", "File not found")
+	} else {
+		role := r.Header.Get("role")
+		idstring := fmt.Sprintf("%d", filemodel.UserID)
+		userid := r.Header.Get("id")
+		if role == "admin" || (role == "user" && idstring == userid) {
+			data, err := upload(r)
+			if err != nil {
+				rsp.Errors.Add("file", err.Error())
+			} else {
+				err := os.Remove(filemodel.Path)
+				if err != nil {
+					rsp.Errors.Add("file", err.Error())
+				}
+				App.DB.Model(&filemodel).Updates(data)
+			}
 		} else {
-			//idstring := fmt.Sprintf("%d", filemodel.UserID)
-			//if idstring != r.Header.Get("id") {
-			//rsp.Errors.Add("ID", "Only owner can change element")
-			//} else {
-			App.DB.Model(&filemodel).Updates(data)
-			//}
+			rsp.Errors.Add("file", "Only owner can change element")
 		}
 	}
 
@@ -198,59 +261,25 @@ func actionDelete(w http.ResponseWriter, r *http.Request) {
 	if file.ID == 0 {
 		rsp.Errors.Add("ID", "File not found")
 	} else {
-		if App.IsTest {
-			App.DB.Unscoped().Delete(&file)
+		role := r.Header.Get("role")
+		idstring := fmt.Sprintf("%d", file.UserID)
+		userid := r.Header.Get("id")
+		if role == "admin" || (role == "user" && idstring == userid) {
+			if App.IsTest {
+				App.DB.Unscoped().Delete(&file)
+			} else {
+				App.DB.Delete(&file)
+			}
+			err := os.Remove(file.Path)
+			if err != nil {
+				rsp.Errors.Add("file", err.Error())
+			}
 		} else {
-			App.DB.Delete(&file)
+			rsp.Errors.Add("file", "Only owner can delete element")
 		}
 	}
 
 	rsp.Data = &file
 
 	w.Write(rsp.Make())
-}
-
-func upload(r *http.Request) (File, error) {
-	// Parse our multipart form, 10 << 20 specifies a maximum
-	// upload of 10 MB files.
-	r.ParseMultipartForm(10 << 20)
-	// FormFile returns the first file for the given key `myFile`
-	// it also returns the FileHeader so we can get the Filename,
-	// the Header and the size of the file
-	file, handler, err := r.FormFile("file")
-	if err != nil {
-		return File{}, errors.New("Error Retrieving the File")
-	}
-	defer file.Close()
-	filename := strings.TrimSuffix(handler.Filename, path.Ext(handler.Filename))
-	fileext := path.Ext(handler.Filename)
-
-	// Create a temporary file within our temp-images directory that follows
-	// a particular naming pattern
-	tempFile, err := ioutil.TempFile("web/uploads", "*"+filename+fileext)
-	if err != nil {
-		return File{}, err
-	}
-	defer tempFile.Close()
-
-	// read all of the contents of our uploaded file into a
-	// byte array
-	fileBytes, err := ioutil.ReadAll(file)
-	if err != nil {
-		return File{}, err
-	}
-	// write this byte array to our temporary file
-	tempFile.Write(fileBytes)
-
-	return File{
-		UserID: 0,
-		Name:   filename,
-		Path:   tempFile.Name(),
-		Ext:    fileext,
-		Preset: "notset",
-		Size:   handler.Size,
-		Status: 0,
-		Type:   0,
-		Hash:   fmt.Sprintf("%x", md5.Sum(fileBytes)),
-	}, nil
 }
